+++ Bundle/AdminBundle/src/DataTable/UserTableType.php
@@ -3,7 +3,6 @@
 namespace Umbrella\AdminBundle\DataTable;
 
 use Doctrine\ORM\QueryBuilder;
-use Symfony\Component\OptionsResolver\OptionsResolver;
 use Symfony\Component\Routing\RouterInterface;
 use Umbrella\AdminBundle\Entity\BaseAdminUser;
 use Umbrella\AdminBundle\UmbrellaAdminConfiguration;
@@ -93,10 +92,5 @@ class UserTableType extends DataTableType
                 }
             }
         ]);
-    }
-
-    public function configureOptions(OptionsResolver $resolver)
-    {
-        $resolver->setDefault('translation_domain', 'UmbrellaAdmin');
     }
 }