+++ Bundle/CoreBundle/src/DataTable/DTO/RowView.php
@@ -6,6 +6,8 @@ class RowView implements \JsonSerializable
 {
     public array $data = [];
 
+    public string $id = '';
+
     public string $class = '';
 
     public array $attr = [];
@@ -13,7 +15,8 @@ class RowView implements \JsonSerializable
     public function jsonSerialize(): array
     {
         $json = $this->data;
-        $json['DT_RowClass'] = $this->class;
+        //$json['DT_RowId'] = $this->id;
+        $json['DT_RowClass'] = 'dt-row ' . $this->class;
         $json['DT_RowAttr'] = $this->attr;
 
         return $json;