+++ Bundle/CoreBundle/src/DataTable/DataTableType.php
@@ -7,10 +7,8 @@ use Symfony\Component\OptionsResolver\OptionsResolver;
 
 class DataTableType
 {
-    public const TAG_SEND_DATA = 'dt:senddata';
-
-    public const DEFAULT_MODE = 'default';
-    public const SELECTION_MODE = 'selection';
+    public const SELECT_MULTIPLE = 'multi';
+    public const SELECT_SINGLE = 'single';
 
     // FIXME : statically called to avoid to have add parent::configureOptions() on all inherit Type class
     final public static function __configureOptions(OptionsResolver $resolver)
@@ -27,6 +25,9 @@ class DataTableType
 
             ->setDefault('table_class', null)
             ->setAllowedTypes('table_class', ['null', 'string'])
+
+            ->setDefault('select', false)
+            ->setAllowedValues('select', [false, self::SELECT_MULTIPLE, self::SELECT_SINGLE])
 
             ->setDefault('paging', function (Options $options) {
                 return !$options['tree'];
@@ -53,12 +54,6 @@ class DataTableType
             })
             ->setAllowedTypes('orderable', 'bool')
 
-            ->setDefault('tree', false)
-            ->setAllowedTypes('tree', 'bool')
-
-            ->setDefault('tree_state', 'collapsed')
-            ->setAllowedValues('tree_state', ['expanded', 'collapsed'])
-
             ->setRequired('dom')
             ->setAllowedTypes('dom', 'string')
 
@@ -66,16 +61,27 @@ class DataTableType
             ->setAllowedTypes('template', 'string');
 
         $resolver
-            ->setDefault('load_url', null)
-            ->setAllowedTypes('load_url', ['string', 'null'])
+            ->setDefault('tree', false)
+            ->setAllowedTypes('tree', 'bool')
 
-            ->setDefault('rowreorder_url', null)
-            ->setAllowedTypes('rowreorder_url', ['string', 'null']);
+            ->setDefault('tree_expanded', false)
+            ->setAllowedTypes('tree_expanded', 'bool');
 
         $resolver
-            ->setDefault('toolbar_class', null)
-            ->setAllowedTypes('toolbar_class', ['string', 'null'])
+            ->setDefault('load_route', null)
+            ->setAllowedTypes('load_route', ['string', 'null'])
 
+            ->setDefault('load_route_params', [])
+            ->setAllowedTypes('load_route_params', 'array');
+
+        $resolver
+            ->setDefault('rowreorder_route', null)
+            ->setAllowedTypes('rowreorder_route', ['string', 'null'])
+
+            ->setDefault('rowreorder_route_params', [])
+            ->setAllowedTypes('rowreorder_route_params', 'array');
+
+        $resolver
             ->setDefault('toolbar_form_name', function (Options $options) {
                 return sprintf('%s_tbf', $options['id']);
             })
@@ -89,7 +95,7 @@ class DataTableType
             ])
             ->setAllowedTypes('toolbar_form_options', 'array')
 
-            ->setDefault('toolbar_template', '@UmbrellaCore/Toolbar/toolbar.html.twig')
+            ->setDefault('toolbar_template', '@UmbrellaCore/DataTable/toolbar.html.twig')
             ->setAllowedTypes('toolbar_template', 'string')
 
             ->setDefault('toolbar_form_data', null);